requirements.txt
requests
yfinance
pytest
pyyaml
flask
python-dotenv
discord.py
pytz
prometheus-client
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
opentelemetry-exporter-prometheus
APScheduler